requirements.txt

gunicorn~=20.1.0
asgiref~=3.5.0
django-environ~=0.8.1
django==4
djangorestframework==3.13.1
djangorestframework-simplejwt~=5.2.0
django-cors-headers~=3.13.0
drf-yasg==1.20.0
requests~=2.27.1
pillow~=9.2.0